refactor(tests): move IncompleteBallisticRuntimeTests to its own file
